moving the game object files to the right folder
